Add a fake project shared to workaround CI crash
TeamCity runs :shared:clean as a first step of the build process.
But now shared is removed from K/N and CI crashes on this step.
At the same time we need this step to build branches with undetached
shared. This patch adds an empty project to workaround this crash.

TODO: Remove this build step and revert this patch when shared
detaching is merged.