[cinterop] Fix mapping of typedefs to pointers to typedefs
Mapping type aliases from stub ir to metadata is kinda tricky.
We need to preserve references to type alias declarations when translating
expanded types. The problem is that we translate C declarations and types independently.
In the most cases these processes are coherent. But not for type aliases to pointers to type aliases.
This commit fixes it.