[Interop][Linker] Add support for metadata-only libraries
We write UniqId of interop declaration in the IdSignature of interop declaration
instead of its mangle when serializing library that depends on interop library.

When deserializing such library, we check IS_NATIVE_INTEROP_LIBRARY flag of IdSignature.
Presence of this flag means that this declaration came from interop library and we should perform
a lookup. It is done in DescriptorByIdSignatureFinder.kt. Right now this lookup is naive and (maybe)
slow.