Support short library names in compiler, interop and platform libs
Fixes for KT-36720 and KT-36721 change a scheme that is used to generate
unique names for klibs: now we use "fully qualified" names that
include a group and a name of a library (e.g. org.jetbrains.kotlin.
native.posix).

But in some cases we still need short library names. For example
to display platform libs in the IDE or to export some symbols of
a library to an ObjC framework.

This patch adds support for such short names to:
 - klib generation in the compiler
 - cinterop tool (both sourcecode and metadata mode)
 - platform libraries (both prebuilt and built on the user side)

Support for C/ObjC export remains.