REVERTME: Force Gradle plugin 1.4.0-dev-8626
Latest changes in the master of the big Kotlin repo are
incompatible with an old Gradle plugin used in the K/N build
and break the composite build. To avoid it, this patch
forces using a compatible Gradle plugin published from a
feature branch.

Note that this patch affects only buildscript dependencies
and doesn't change a frontend version the K/N compiler is
built against.

This patch must be reverted during the compiler update.