Use a system property to lock in the dependencies downloader
Current implementation of synchronization in the dependencies
downloader works uses a static mutex and a file lock. It
doesn't work if compiler instances are started in the same
process but in different classloaders (see KT-39450).

This patch workes around this problem by using a system
property instead of a static mutex for synchronization.

Issue #KT-39450 fixed