[Linkage][Lowering] InternalABI
To effectively access enum entries during separate compilation (in other words, when compiling with caches), we need to reference
Enum.$OBJECT.get-VALUES function. It is not a part of a public API because it is generated during EnumClassLowering. What are possible solutions?
1. Lazy IR. It is no-go, because there is no descriptor for it.
2. Generate IrFiles and appropriate packages during lowering. We can't do it, because lowerings are file-based and ConcurrentModificationException will occur.
2.1. Store declarations in list and add file later? Still not an option, because .parent (or module) for these declaration won't be initialized which breaks lowerings.
3. Something ugly. Here we go! We add a special file before lowerings which stores all declaration that form an internal ABI.

(cherry picked from commit 31bad64728c5490969000a8e451165f9079761bc)