[build] Fix outputs for the cache task
A klib filename without extension was used to determine a name
of a cache directory for this klib. This is incorrect if the
klib is stored in a directory with dots in its name. This breaks
up-to-date checks, so Gradle doesn't rerun the task if the
cache directory was deleted.

This patch fixes this issue.