[build] Fix stdlib caches in the light distribution
Latest changes in the cache layout has added a new bitcode_deps file
to a cache directory. But a task building a light distribution
bundle doesn't handle this. So such files for stdlib and
kotlinx.cli are missing in the light distribution. This causes
failures when building platform libs and caches at the user side.

This patch fixes this issue by taking the new cache layout into
account in the bundle task.